backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2image
pytesseract
google-generativeai
python-dotenv
reportlab
pymupdf
httpx
redis
sqlalchemy
psycopg[binary]
alembic
